+++ src/ultimate_notion/database.py
@@ -14,7 +14,7 @@ from typing_extensions import Self, TypeVar
 
 from ultimate_notion.blocks import ChildrenMixin, DataObject, wrap_icon
 from ultimate_notion.core import NotionEntity, WorkspaceType, get_active_session, get_repr, resolve_ref
-from ultimate_notion.emoji import CustomEmoji, Emoji
+from ultimate_notion.emoji import BuiltInIcon, CustomEmoji, Emoji
 from ultimate_notion.errors import ReadOnlyPropertyError, SchemaError, SListError, UnsetError
 from ultimate_notion.file import AnyFile
 from ultimate_notion.obj_api import blocks as obj_blocks
@@ -73,7 +73,7 @@ class DataContainer(DataObject[DC_co], wraps=obj_blocks.DataObject):
         session.api.data_sources.update(self.obj_ref, description=text.obj_ref)
 
     @property
-    def icon(self) -> AnyFile | Emoji | CustomEmoji | None:
+    def icon(self) -> AnyFile | Emoji | CustomEmoji | BuiltInIcon | None:
         """Return the icon of this data source as file or emoji."""
         if (icon := self.obj_ref.icon) is None:
             return None